+++ kernel/src/syscalls/hardware.rs
@@ -1,3 +1,3 @@
-pub fn change_video_mode(mode: u32) {
+pub fn change_video_mode(mode: u8) {
   crate::hardware::vga::driver::request_mode_change(mode);
 }
